Reveal flag in enrichSystemProperties()
The enrichSystemProperties() function used a hidden flag to detect
whether the System properties have already been enriched, in which case
it will abort with an IllegalStateException.
That flag has now been revealed as the System property "util.enriched",
which is set to `true` if the System properties have already been
enriched (and consequently can be set to `false` or just deleted to run
the enrichment process again).