Improve modularization of human interrupts
The way human interrupts were created and shown was not really
object-oriented, or even modular. The TwentyApp (another unnecessary
class to be removed in a future commit) was creating and managing the
InterruptWindow, while the Twenty object, for its "windowTest" routine,
was creating an additional instance of InterruptWindow.
To fix these issues, a new class – HumanInterrupter – now encapsulates,
hides and manages a private instance of InterruptWindow. The TwentyApp
now holds a reference to that HumanInterrupter, and uses its
interruptHuman() method to interrupt the human user.
Additionally, HumanInterrupter offers a static method, testWindow(),
which is utilized by the Twenty object for window testing.