Improve InfoWindow display updating mechanism
The InfoWindow no longer uses a separate Thread to update the screen
time information, and instead relies on user interactions as impulse
givers (with a minimum elapsed time to prevent unnecessary updates).
This transition, which saves resources by basically using the event
thread to update the screen time display, was possible because the
screen time is now shown in minutes, and no longer in seconds (so the
user doesn't notice the more infrequent, irregular display updates).

Since the InfoWindow no longer requires a separate animation Thread,
i.e. no longer is an "animated window", it does not need to extend
AnimatedJFrame, and is now a regular subclass of JFrame.