Merge TwentyApp into Twenty
Due to the recent improvements in the program's structures, a
separation between the Twenty main object and a TwentyApp is no longer
necessary, so the two have been merged:

Twenty is now a class (not an object), which behaves like the
now-obsolete and removed TwentyApp, and which is instantiated in the
Twenty companion object (where the static main method is now located).
The Twenty companion object now provides all the functionality
previously provided by the Twenty object, completing the merger.