Add sections
Sections are a quick way to isolate use-once subroutines (which are
only called once, and only by their surrounding routines, i.e. no code
is / can be reused in them) from their surrounding code, resulting in a
more structured program without creating lots of private methods which
are only used to isolate parts of one code path from each other.

Code sections are implemented in Kotlin with an inline function, which
takes a name and a closure block with no arguments returning Unit.
The String name is optimized away at compile time and the function is
inlined, leaving no trace of sectioning in the compiled output.