Change how Resources work
Resources is no longer a static object in the main twenty package, but
instead an instance of a class created by Twenty and then passed to all
the instances of UI classes which Twenty creates.

All resources are now located in the sub-package ".../twenty/res/", and
icons have been moved to ".../twenty/res/icons/".

SettingsIcon has been deleted because it is no longer required.