Make popupMenu an extension function of TrayIcon
In Fusion, free (non-extension) functions which initiate the builder
pattern should be avoided, since it is unclear whether they are already
attached or need to be attached manually. This was exactly the case
with the popupMenu(String, () -> Unit) function, which initiated a
builder pattern for an AWT PopupMenu but did not attach it to any
component (since it was free-form, and not an extension function).
To avoid this counter-intuitive behaviour (all other functions of this
scheme automatically attach themselves), the function has been
transformed into an extension of TrayIcon, which now automatically
attaches the generated PopupMenu by assigning it to the TrayIcon's
`popupMenu` property.
Similar popupMenu() extensions for other types of components will not
be created unless required, per YAGNI.