Remove animations from HumanInterrupter
Animations had been removed because they were not performant on
platforms other than Windows and had a flawed implementation.
Now, InterruptWindow extends JFrame instead of AnimatedJFrame,
and the auto-closing after a specified duration (in seconds) is handled
by a Swing Timer.
This reduces the overall complexity and increases both performance and
compliance (with Swing).