Add Java plugin and JetBrains annotations library
The Java plugin will be required because implementations of Swing UI
classes require overriding a static method (which is not supported in
Kotlin right now), so Java will have to be used for those classes.

JetBrains annotations are used for Java/Kotlin interoperability (mainly
null safety).

The JVM target has been upgraded to 11 for both Java and Kotlin.