Create MetalUtils
The original MetalUtils used in the Metal L&F are, regrettably,
package-private, which is why they need to be imitated in order to
implement custom UI elements which are compatible with Metal's L&F.

Note that the functionality is only imitated; the actual L&F components
may still be dissimilar to "true" Metal L&F components. The goal is
consistency/coherence, not emulation.