Add Makefile
A Makefile is used for the install/uninstall tasks because Gradle is
unfamiliar to many non-JVM programmers and users of POSIX systems, and
its documentation isn't adequate for quick configuration / tweaking on
potential target platforms.
Makefiles are more familiar, have a better learning curve,
and are, for lack of a better description, more POSIX-compliant.