Upgrade dependencies and overhaul app integrity
Upgrade dependencies SystemTray and Utilities.

Because the automatic shutdown hook installation procedure has been
removed from SystemTray, a shutdown hook has to be installed manually.
A new object called AppIntegrity encapsulates this operation, and also
includes a shutdown() and a displayThrowable() procedure.

Some parts of the codebase had to be adjusted to the changed API of the
Utilities library.